Klasse FuzzyModel aangemaakt ViewController vraagt nu ES-waarden van FuzzyModel

+++ CLESMonitor/CLESMonitor/Model/ESModel.cs
@@ -12,6 +12,6 @@ namespace CLESMonitor.Model
         /// Bereken opnieuw de model-waarde
         /// </summary>
         /// <returns>De model-waarde</returns>
-        public abstract int calculateModelValue();
+        public abstract double calculateModelValue();
     }
 }
